tgui: custom font bluescreen fix, typescript'd (#84601)
## About The Pull Request
Issue was caused by the fact that it was forwarding the custom font
_event_ rather than the string value. Caused by #82652. onChange is
`event, value`. See the problem here:
https://github.com/tgstation/tgstation/blob/c9e785ce3e825edb16d2b560f79d568e3fbe4fc6/tgui/packages/tgui-panel/settings/SettingsPanel.jsx#L143-L149

While poking around I figured now's the best time to make this thing
TypeScript. So the entire chat settings ui has been refactored, cleaned
up, typescripted. Functionality remains the same.
## Why It's Good For The Game
Fixes #84593

![Screenshot 2024-07-02
165624](https://github.com/tgstation/tgstation/assets/42397676/d41d8635-b8c6-4f85-9aca-d2cd2a973325)

## Changelog
:cl:
fix: Fixed the bluescreen while setting a custom font in tgui panel
/:cl:
