+++ tgui/packages/tgui-panel/settings/SettingsGeneral.tsx
@@ -1,0 +1,173 @@
+import { toFixed } from 'common/math';
+import { capitalize } from 'common/string';
+import { useState } from 'react';
+import { useDispatch, useSelector } from 'tgui/backend';
+import {
+  Button,
+  Collapsible,
+  Divider,
+  Input,
+  LabeledList,
+  NumberInput,
+  Section,
+  Stack,
+} from 'tgui/components';
+
+import { clearChat, saveChatToDisk } from '../chat/actions';
+import { THEMES } from '../themes';
+import { updateSettings } from './actions';
+import { FONTS } from './constants';
+import { selectSettings } from './selectors';
+
+export function SettingsGeneral(props) {
+  const { theme, fontFamily, fontSize, lineHeight } =
+    useSelector(selectSettings);
+  const dispatch = useDispatch();
+  const [freeFont, setFreeFont] = useState(false);
+
+  return (
+    <Section>
+      <LabeledList>
+        <LabeledList.Item label="Theme">
+          {THEMES.map((THEME) => (
+            <Button
+              key={THEME}
+              selected={theme === THEME}
+              color="transparent"
+              onClick={() =>
+                dispatch(
+                  updateSettings({
+                    theme: THEME,
+                  }),
+                )
+              }
+            >
+              {capitalize(THEME)}
+            </Button>
+          ))}
+        </LabeledList.Item>
+        <LabeledList.Item label="Font style">
+          <Stack.Item>
+            {!freeFont ? (
+              <Collapsible
+                title={fontFamily}
+                width={'100%'}
+                buttons={
+                  <Button
+                    icon={freeFont ? 'lock-open' : 'lock'}
+                    color={freeFont ? 'good' : 'bad'}
+                    onClick={() => {
+                      setFreeFont(!freeFont);
+                    }}
+                  >
+                    Custom font
+                  </Button>
+                }
+              >
+                {FONTS.map((FONT) => (
+                  <Button
+                    key={FONT}
+                    fontFamily={FONT}
+                    selected={fontFamily === FONT}
+                    color="transparent"
+                    onClick={() =>
+                      dispatch(
+                        updateSettings({
+                          fontFamily: FONT,
+                        }),
+                      )
+                    }
+                  >
+                    {FONT}
+                  </Button>
+                ))}
+              </Collapsible>
+            ) : (
+              <Stack>
+                <Input
+                  width={'100%'}
+                  value={fontFamily}
+                  onChange={(e, value) =>
+                    dispatch(
+                      updateSettings({
+                        fontFamily: value,
+                      }),
+                    )
+                  }
+                />
+                <Button
+                  ml={0.5}
+                  icon={freeFont ? 'lock-open' : 'lock'}
+                  color={freeFont ? 'good' : 'bad'}
+                  onClick={() => {
+                    setFreeFont(!freeFont);
+                  }}
+                >
+                  Custom font
+                </Button>
+              </Stack>
+            )}
+          </Stack.Item>
+        </LabeledList.Item>
+        <LabeledList.Item label="Font size">
+          <NumberInput
+            width="4.2em"
+            step={1}
+            stepPixelSize={10}
+            minValue={8}
+            maxValue={32}
+            value={fontSize}
+            unit="px"
+            format={(value) => toFixed(value)}
+            onChange={(value) =>
+              dispatch(
+                updateSettings({
+                  fontSize: value,
+                }),
+              )
+            }
+          />
+        </LabeledList.Item>
+        <LabeledList.Item label="Line height">
+          <NumberInput
+            width="4.2em"
+            step={0.01}
+            stepPixelSize={2}
+            minValue={0.8}
+            maxValue={5}
+            value={lineHeight}
+            format={(value) => toFixed(value, 2)}
+            onDrag={(value) =>
+              dispatch(
+                updateSettings({
+                  lineHeight: value,
+                }),
+              )
+            }
+          />
+        </LabeledList.Item>
+      </LabeledList>
+      <Divider />
+      <Stack fill>
+        <Stack.Item grow mt={0.15}>
+          <Button
+            icon="save"
+            tooltip="Export current tab history into HTML file"
+            onClick={() => dispatch(saveChatToDisk())}
+          >
+            Save chat log
+          </Button>
+        </Stack.Item>
+        <Stack.Item mt={0.15}>
+          <Button.Confirm
+            icon="trash"
+            tooltip="Erase current tab history"
+            onClick={() => dispatch(clearChat())}
+          >
+            Clear chat
+          </Button.Confirm>
+        </Stack.Item>
+      </Stack>
+    </Section>
+  );
+}
